nr_decoders_fix + batchsize optimization

+++ ensamble/train.py
@@ -71,7 +71,7 @@ def get_arg_parser():
     parser.add_argument("--TRANSFORM_IMAGE", type= list, default=TRANSFORM_IMAGE, help="Image transformation")
     parser.add_argument("--TRANSFORM_MASK", type= list, default=TRANSFORM_MASK, help="Mask transformation")
     
-    parser.add_argument("--nr_of_decoders", type=float, default=6, help="Dropout rate for the model")
+    parser.add_argument("--nr_of_decoders", type=int, default=6, help="Dropout rate for the model")
     parser.add_argument("--wandb_mode", type=str, default="offline", help="Wandb mode")
